+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Changelog
 ---
 ## 0.8.3
+* Fixed that local anchors are not ignored.
 * Better scrollbar styles for WebKit/Blink browsers in Windows.
 
 ## 0.8.2